requirements.txt
streamlit
gtts
google-generativeai